requirements.txt
Django==4.0.3
Pygments==2.11.2
djangorestframework~=3.13.1